Add policy for document

+++ app/controllers/documents_controller.rb
@@ -3,11 +3,13 @@ class DocumentsController < ApplicationController
 
   def index
     @document = Document.new
+    @documents = policy_scope(@funeral.documents)
   end
 
   def create
     @document = @funeral.documents.build(document_params)
     @document.file_name = params[:document][:file].original_filename
+    authorize @document
     if @document.save
       redirect_to funeral_documents_path(@funeral), notice: 'Document uploaded!'
     else
@@ -17,6 +19,7 @@ class DocumentsController < ApplicationController
 
   def destroy
     document = @funeral.documents.find(params[:id])
+    authorize document
     document.destroy
     redirect_to funeral_documents_path(@funeral), notice: 'Document deleted!'
   end
